New file called "RPG_remake.py"
This file is basically a rewrite of all of the code because jesus christ that was a bloody mess. What this file has for now is just structure generation, but this time it's WAAAY more simple and automatic. instead of adding every structure individually now there's a list with the structures and it gets the random structure based on that list.